env handled in main scripts
